# openapi.yaml
openapi: 3.0.3
info:
  title: AI 기술 면접 연습 플랫폼 API
  version: 0.4.1
  description: 핵심 사용자 흐름과 최신 요구사항 변경안을 반영한 OpenAPI 계약. API 변경 시 docs/requirements.md,
    docs/project/open-items.md와 함께 수정한다. 핵심 성공 응답은 success/data/meta 형태와 예시 응답을 포함한다.
servers:
- url: /api/v1
...
components:
  securitySchemes:
    AuthorizationHeader:
      type: apiKey
      in: header
      name: Authorization
      description: '형식: Bearer {apiKey} {accessToken}'
  schemas:
    Meta:
      type: object
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        pagination:
          $ref: '#/components/schemas/Pagination'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            fieldErrors:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
          required:
          - code
          - message
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - success
      - error
      - meta
...
    Document:
      type: object
      properties:
        id:
          type: integer
          format: int64
        documentType:
          type: string
          enum:
          - resume
          - award
          - certificate
          - other
        originalFileName:
          type: string
        mimeType:
          type: string
          description: '허용 MIME type: application/pdf, application/vnd.openxmlformats-officedocument.wordprocessingml.document,
            text/markdown'
        fileSizeBytes:
          type: integer
          format: int64
          description: 최대 10485760 bytes (10MB)
        extractStatus:
          type: string
          enum:
          - pending
          - success
          - failed
        uploadedAt:
          type: string
          format: date-time
        extractedAt:
          type: string
          format: date-time
          nullable: true
        extractedText:
          type: string
          nullable: true
      required:
      - id
      - documentType
      - originalFileName
      - mimeType
      - fileSizeBytes
      - extractStatus
      - uploadedAt
...
    InterviewQuestion:
      type: object
      properties:
        id:
          type: integer
          format: int64
        questionOrder:
          type: integer
        questionType:
          type: string
          enum:
          - experience
          - project
          - technical_cs
          - technical_stack
          - behavioral
          - follow_up
        difficultyLevel:
          type: string
          enum:
          - easy
          - medium
          - hard
        questionText:
          type: string
        parentQuestionId:
          type: integer
          format: int64
          nullable: true
        sourceApplicationQuestionId:
          type: integer
          format: int64
          nullable: true
      required:
      - id
      - questionOrder
      - questionType
      - difficultyLevel
      - questionText
...
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
      required:
      - field
      - reason
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 1
      required:
      - page
      - size
      - totalElements
      - totalPages
...
    InterviewQuestionSetSummary:
      type: object
      properties:
        questionSetId:
          type: integer
          format: int64
        applicationId:
          type: integer
          format: int64
        title:
          type: string
        questionCount:
          type: integer
        difficultyLevel:
          type: string
          enum:
          - easy
          - medium
          - hard
        createdAt:
          type: string
          format: date-time
      required:
      - questionSetId
      - applicationId
      - title
      - questionCount
      - difficultyLevel
      - createdAt
...
security:
- AuthorizationHeader: []
paths:
...
  /documents:
    post:
      tags:
      - Documents
      summary: 문서 업로드
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                documentType:
                  type: string
                  enum:
                  - resume
                  - award
                  - certificate
                  - other
                  description: resume | award | certificate | other
                file:
                  type: string
                  format: binary
                  description: '허용 형식: PDF, DOCX, MD / 최대 10MB / 기본 최대 5개'
              required:
              - documentType
              - file
      responses:
        '201':
          description: 업로드 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Document'
                  meta:
                    $ref: '#/components/schemas/Meta'
                required:
                - success
                - data
                - meta
              examples:
                documentUploadSuccess:
                  value:
                    success: true
                    data:
                      id: 301
                      documentType: resume
                      originalFileName: resume_2026.pdf
                      mimeType: application/pdf
                      fileSizeBytes: 834221
                      extractStatus: pending
                      uploadedAt: '2026-03-17T10:31:10Z'
                      extractedAt: null
                      extractedText: null
                    meta: *id001
        '400':
          description: 지원하지 않는 파일 형식 또는 최대 용량 초과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                documentTooLargeError:
                  value:
                    success: false
                    error:
                      code: DOCUMENT_FILE_TOO_LARGE
                      message: 파일 크기는 10MB를 초과할 수 없습니다.
                      retryable: false
                      fieldErrors:
                      - field: file
                        reason: max_10mb
                    meta: *id001
      description: 이력서, 수상기록 등 문서를 업로드하고 텍스트 추출을 시작한다. 허용 형식은 PDF, DOCX, MD이고 파일 1개당
        최대 10MB다. 기본 업로드 수는 5개다.
    get:
      tags:
      - Documents
      summary: 문서 목록 조회
      responses:
        '200':
          description: 문서 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  meta:
                    $ref: '#/components/schemas/Meta'
                required:
                - success
                - data
                - meta
              examples:
                documentListSuccess:
                  value:
                    success: true
                    data:
                    - &id003
                      id: 301
                      documentType: resume
                      originalFileName: resume_2026.pdf
                      mimeType: application/pdf
                      fileSizeBytes: 834221
                      extractStatus: success
                      uploadedAt: '2026-03-17T10:31:10Z'
                      extractedAt: '2026-03-17T10:31:12Z'
                      extractedText: Spring Boot 기반 프로젝트 경험 ...
                    meta: *id002
  /documents/{documentId}:
    get:
      tags:
      - Documents
      summary: 문서 상세 조회
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: 문서 상세
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Document'
                  meta:
                    $ref: '#/components/schemas/Meta'
                required:
                - success
                - data
                - meta
              examples:
                documentDetailSuccess:
                  value:
                    success: true
                    data: *id003
                    meta: *id001
    delete:
      tags:
      - Documents
      summary: 문서 삭제
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: 삭제 성공
        '409':
          description: 참조 중 문서
...
  /interview/question-sets:
    post:
      tags:
      - Interview
      summary: 면접 질문 세트 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                applicationId:
                  type: integer
                  format: int64
                title:
                  type: string
                questionCount:
                  type: integer
                  minimum: 1
                  maximum: 20
                  description: 1회 최대 20개
                difficultyLevel:
                  type: string
                  enum:
                  - easy
                  - medium
                  - hard
                questionTypes:
                  type: array
                  items:
                    type: string
                    enum:
                    - experience
                    - project
                    - technical_cs
                    - technical_stack
                    - behavioral
                    - follow_up
              required:
              - applicationId
              - questionCount
              - difficultyLevel
              - questionTypes
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/InterviewQuestionSetSummary'
                  meta:
                    $ref: '#/components/schemas/Meta'
                required:
                - success
                - data
                - meta
              examples:
                questionSetCreateSuccess:
                  value:
                    success: true
                    data:
                      questionSetId: 701
                      applicationId: 401
                      title: 네이버 백엔드 예상 질문 세트
                      questionCount: 5
                      difficultyLevel: medium
                      createdAt: '2026-03-17T10:34:00Z'
                    meta: *id001
      description: 지원 단위를 기준으로 면접 질문 세트를 생성한다. 서버는 질문 수 불일치, schema violation, malformed
        JSON에 대해 최대 2회 재시도한다. 질문 수는 최대 20개다.
    get:
      tags:
      - Interview
      summary: 면접 질문 세트 목록 조회
      responses:
        '200':
          description: 질문 세트 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InterviewQuestionSetSummary'
                  meta:
                    $ref: '#/components/schemas/Meta'
                required:
                - success
                - data
                - meta
              examples:
                questionSetListSuccess:
                  value:
                    success: true
                    data:
                    - questionSetId: 701
                      applicationId: 401
                      title: 네이버 백엔드 예상 질문 세트
                      questionCount: 5
                      difficultyLevel: medium
                      createdAt: '2026-03-17T10:34:00Z'
                    meta: *id002
...
  /interview/question-sets/{questionSetId}/questions:
    post:
      tags:
      - Interview
      summary: 면접 질문 세트에 수동 질문 추가
      parameters:
      - name: questionSetId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                questionText:
                  type: string
                questionType:
                  type: string
                  enum:
                  - experience
                  - project
                  - technical_cs
                  - technical_stack
                  - behavioral
                  description: 1차 범위에서는 독립 질문만 허용하며 follow_up은 받지 않는다.
                difficultyLevel:
                  type: string
                  enum:
                  - easy
                  - medium
                  - hard
              required:
              - questionText
              - questionType
              - difficultyLevel
      responses:
        '201':
          description: 수동 질문 추가 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/InterviewQuestion'
                  meta:
                    $ref: '#/components/schemas/Meta'
                required:
                - success
                - data
                - meta
              examples:
                manualQuestionAddSuccess:
                  value:
                    success: true
                    data:
                      id: 803
                      questionOrder: 3
                      questionType: behavioral
                      difficultyLevel: medium
                      questionText: 이 회사에 지원한 이유를 설명해주세요.
                      parentQuestionId: null
                      sourceApplicationQuestionId: null
                    meta: *id001
        '400':
          description: 요청 검증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidManualQuestionRequest:
                  value:
                    success: false
                    error:
                      code: REQUEST_VALIDATION_FAILED
                      message: 요청 값을 확인해주세요.
                      retryable: false
                      fieldErrors:
                      - field: questionText
                        reason: blank
                    meta: *id001
        '404':
          description: 질문 세트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                questionSetNotFound:
                  value:
                    success: false
                    error:
                      code: RESOURCE_NOT_FOUND
                      message: 질문 세트를 찾을 수 없습니다.
                      retryable: false
                    meta: *id001
        '409':
          description: 이미 시작된 질문 세트는 편집할 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                questionSetNotEditable:
                  value:
                    success: false
                    error:
                      code: INTERVIEW_QUESTION_SET_NOT_EDITABLE
                      message: 이미 면접이 시작된 질문 세트는 수정할 수 없습니다.
                      retryable: false
                    meta: *id001
      description: 세션 시작 전 질문 세트에 수동 질문을 추가한다. questionOrder는 요청으로 받지 않으며 서버가 마지막 순번 다음 값으로
        자동 부여한다. 1차 범위에서는 독립 질문만 허용하고 parentQuestionId는 받지 않는다.
  /interview/question-sets/{questionSetId}/questions/{questionId}:
    delete:
      tags:
      - Interview
      summary: 면접 질문 세트에서 질문 개별 삭제
      parameters:
      - name: questionSetId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: questionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '400':
          description: 마지막 남은 질문은 삭제할 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                lastQuestionDeleteRejected:
                  value:
                    success: false
                    error:
                      code: REQUEST_VALIDATION_FAILED
                      message: 질문 세트는 최소 1개 질문을 유지해야 합니다.
                      retryable: false
                      fieldErrors:
                      - field: questionId
                        reason: min_1
                    meta: *id001
        '404':
          description: 질문 세트 또는 질문을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                interviewQuestionNotFound:
                  value:
                    success: false
                    error:
                      code: RESOURCE_NOT_FOUND
                      message: 질문을 찾을 수 없습니다.
                      retryable: false
                    meta: *id001
        '409':
          description: 이미 시작된 질문 세트는 편집할 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                deleteQuestionSetNotEditable:
                  value:
                    success: false
                    error:
                      code: INTERVIEW_QUESTION_SET_NOT_EDITABLE
                      message: 이미 면접이 시작된 질문 세트는 수정할 수 없습니다.
                      retryable: false
                    meta: *id001
        '204':
          description: 질문 삭제 성공
      description: 세션 시작 전 질문 세트에서 질문을 hard delete 한다. 삭제 후 남은 질문의 questionOrder는 서버가 빈 번호 없이
        1..N 형태로 다시 정렬한다. 질문 세트는 최소 1개 질문을 유지해야 하므로 마지막 남은 질문 삭제는 허용하지 않는다.
...